# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microcredit management system API for Argentina",
        "title": "Crédito Villanueva API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@creditovillanueva.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/loans/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of loans awaiting approval",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Loans"
                ],
                "summary": "List pending loans",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/loans/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Approves a pending loan request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Loans"
                ],
                "summary": "Approve a loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/loans/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancels a loan that has not yet been fully disbursed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Loans"
                ],
                "summary": "Cancel a loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/loans/{id}/disburse": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Disburses an approved loan, crediting the funds to the client's account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Loans"
                ],
                "summary": "Disburse a loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/loans/{id}/prepay": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Processes an early cancellation or partial prepayment of a loan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Loans"
                ],
                "summary": "Early loan prepayment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Prepayment amount",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PrepayLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/loans/simulate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Calculates an amortization schedule for a given loan amount, installments, and type (French or German)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loans"
                ],
                "summary": "Simulate a loan",
                "parameters": [
                    {
                        "description": "Simulation parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SimulateLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SimulationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/me/loans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of loans belonging to the authenticated client",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loans"
                ],
                "summary": "List my loans",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submits a loan request against an approved credit line for the authenticated client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loans"
                ],
                "summary": "Request a new loan",
                "parameters": [
                    {
                        "description": "Loan request data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequestLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.LoanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/me/loans/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the full details of a specific loan including its installment schedule",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loans"
                ],
                "summary": "Get loan details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.InstallmentResponse": {
            "type": "object",
            "properties": {
                "capitalAmount": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "interestAmount": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "paidAmount": {
                    "type": "string"
                },
                "paidAt": {
                    "type": "string"
                },
                "remainingAmount": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "totalAmount": {
                    "type": "string"
                }
            }
        },
...
        "dto.LoanResponse": {
            "type": "object",
            "properties": {
                "amortizationType": {
                    "type": "string"
                },
                "clientId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creditLineId": {
                    "type": "string"
                },
                "disbursedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "installments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.InstallmentResponse"
                    }
                },
                "interestRate": {
                    "type": "string"
                },
                "numInstallments": {
                    "type": "integer"
                },
                "principal": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "totalPaid": {
                    "type": "string"
                },
                "totalRemaining": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PrepayLoanRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                }
            }
        },
...
        "dto.RequestLoanRequest": {
            "type": "object",
            "required": [
                "amortizationType",
                "amount",
                "creditLineId",
                "numInstallments"
            ],
            "properties": {
                "amortizationType": {
                    "type": "string",
                    "enum": [
                        "french",
                        "german"
                    ]
                },
                "amount": {
                    "type": "string"
                },
                "creditLineId": {
                    "type": "string"
                },
                "numInstallments": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "dto.SimulateLoanRequest": {
            "type": "object",
            "required": [
                "amortizationType",
                "amount",
                "creditLineId",
                "numInstallments"
            ],
            "properties": {
                "amortizationType": {
                    "type": "string",
                    "enum": [
                        "french",
                        "german"
                    ]
                },
                "amount": {
                    "type": "string"
                },
                "creditLineId": {
                    "type": "string"
                },
                "numInstallments": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "dto.SimulationResponse": {
            "type": "object",
            "properties": {
                "installments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.InstallmentResponse"
                    }
                },
                "interestRate": {
                    "type": "string"
                },
                "principal": {
                    "type": "string"
                },
                "totalInterest": {
                    "type": "string"
                },
                "totalPayment": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}